The new jnect emfstore and replay features now use extension points.
The core plugin has a new extension point for retrieving a recording body
model. Its state is now back to the one from the jnect repository otherwise.

The gef plugin has the same state as in the jnect repository, except for
one null-pointer check. This needs to be verified and might also get
removed.

Added a replay plugin responsible for the replaying. Only supports full
replay so far. This plugin offers an extension point, where model stores
can extend (currently the jnect.emfstore as the only one).

Last but not least added a new project containing code formatters. These
have been aplied to the projects in this repository.